feat: move DataBuffer implementations for testing to helper module
