+++ tests/pamiq_core/data/test_container.py
@@ -1,23 +1,9 @@
 import pytest
 
-from pamiq_core.data.buffer import BufferData, DataBuffer, StepData
 from pamiq_core.data.container import DataCollectorsDict, DataUsersDict
 from pamiq_core.data.interface import DataCollector, DataUser
 
-
-class MockDataBuffer(DataBuffer):
-    def __init__(self, collecting_data_names: list[str], max_size: int) -> None:
-        super().__init__(collecting_data_names, max_size)
-        self.data: list[StepData] = []
-
-    def add(self, step_data: StepData) -> None:
-        if len(self.data) < self.max_size:
-            self.data.append(step_data)
-
-    def get_data(self) -> BufferData:
-        return {
-            name: [d[name] for d in self.data] for name in self._collecting_data_names
-        }
+from .helpers import MockDataBuffer
 
 
 class TestDataUsersDict:
